Moved similar files into user folder
