+++ routes/users/profile.js
@@ -1,0 +1,8 @@
+const express = require("express");
+const router = express.Router();
+
+const profileController = require("../../controllers/users/profile");
+
+router.get("/", profileController.profile);
+
+module.exports = router;
