+++ routes/users/register.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
 
-const controller = require("../controllers/register");
+const controller = require("../../controllers/users/register");
 
 router.post("/", controller.register);
