+++ security/authorization.middleware.js
@@ -9,10 +9,20 @@ module.exports = async (req, res, next) => {
   const token = req.header("auth");
 
   if (token) {
-    const decoded = await jwt.verify(token, JWT_KEY);
-    res.locals.token = token;
-    next();
-    return;
+    try {
+      const decoded = await jwt.verify(token, JWT_KEY);
+      res.locals.token = token;
+      next();
+      return;
+    } catch (err) {
+      console.error("Invalid JWT token");
+      res
+        .status(401)
+        .send(
+          "You must be logged in to see this page! Please send a authorization token!"
+        );
+      return;
+    }
   }
 
   res
